consolidate into banners directory
