+++ .gitignore
@@ -39,3 +39,7 @@ yarn-error.log*
 
 # typescript
 *.tsbuildinfo
+docs/**/*
+
+#allow
+!docs/**/*.md
