+++ src/lib/context/getGlobalElements.ts
@@ -1,5 +1,5 @@
 import { GetServerSidePropsContext, GetStaticPropsContext } from 'next'
-import { transformBannerDataService } from '@/templates/globals/banner/dataService'
+import { transformBannerDataService } from '@/templates/globals/banners/banner/dataService'
 import { drupalClient } from '@/lib/utils/drupalClient'
 import { LayoutProps } from '@/templates/globals/wrapper'
