+++ src/templates/globals/banners/banner/nodeBanner.json
