+++ src/templates/globals/banners/promoBanner/index.tsx
