+++ src/templates/globals/banners/promoBanner/nodePromoBanner.json
